Set label text color based on background color
Summary
--

This is straight from ChatGPT, but it seems to work well, and the results mostly
match rainbow-mode in Emacs.

It's slightly annoying that you can't use non-hex colors now when before you
could use any CSS color, but this is probably the most reasonable approach.

Test Plan
--

New tests on the three main label colors in my config