Add pull-to-refresh
Summary
--

This PR fixes the lack of refresh in the bookmarked/saved-to-home-screen version
of the site on iOS (not sure if Android is also affected).

This feels a bit ugly and is straight out of GPT, but it seems to work.

Test Plan
--

Manual testing on my phone and in the browser